+++ services/prompter.js
@@ -71,23 +71,7 @@ async function Prompter(program, requests) {
     }
   }
 
-  if (isRequested('dbSchema')) {
-    if (process.env.DATABASE_SCHEMA) {
-      envConfig.dbSchema = process.env.DATABASE_SCHEMA;
-    } else {
-      prompts.push({
-        type: 'input',
-        name: 'dbSchema',
-        message: 'What\'s the database schema? [optional]',
-        description: 'Leave blank by default',
-        when: answers => answers.dbDialect !== 'sqlite' && answers.dbDialect !== 'mongodb',
-        default: (args) => {
-          if (args.dbDialect === 'postgres') { return 'public'; }
-          return '';
-        },
-      });
-    }
-  }
+  envConfig.dbSchema = program.schema || process.env.DATABASE_SCHEMA || '';
 
   if (isRequested('dbHostname')) {
     if (process.env.DATABASE_HOST) {