+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Command Generate - Add  "Application port" parameter (`-p, --application-port`) to `lumber generate`.
 - Command Generate - Add  "Application host" parameter (`-H, --application-host`) to `lumber generate`.
 - Command Generate - Add  "SSL" parameter (`-S, --ssl`) to `lumber generate`.
+- Command generate - Add "Schema" parameter (`-s, --schema`) to `lumber generate`.
 
 ### Fixed
 - Command Generate - Avoid foreignKey to conflict with relationship.